% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data
Get BAAC data in a R format}
\usage{
get_data(years = 2018)
}
\arguments{
\item{years}{Vector of years to download, by default 2018}
}
\value{
A dataframe containing the BAAC data for requested years
}
\description{
get_data
Get BAAC data in a R format
}
\examples{
get_data(2018)
get_data(2016:2018)
}
