% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_raw_datasets}
\alias{get_raw_datasets}
\title{Get BAAC csv files from data.gouv.fr to a specified folder}
\usage{
get_raw_datasets(years, path)
}
\arguments{
\item{years}{Vector of years to download}

\item{path}{Folder where to put the csv files}
}
\value{
File list and path to them
}
\description{
Get BAAC csv files from data.gouv.fr to a specified folder
}
\examples{
get_raw_datasets(2018, tempdir())
get_raw_datasets(2016:2018, tempdir())
}
