% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{get_datasets
Get BAAC datasets from data.gouv.fr}
\usage{
get_datasets(year = NA)
}
\arguments{
\item{year}{Vector of years to download}
}
\value{
List of all available datasets if empty or the aggregated datasets for a specific year
}
\description{
get_datasets
Get BAAC datasets from data.gouv.fr
}
\examples{
get_datasets()
get_datasets(2018)
get_datasets(2016:2018)
}
